// src/app/shared/models/patient.model.ts
export interface Patient {
  firstname: string;
  lastname: string;
  patient_id: number;
  heart_rate: string;
  dia_blood_pressure: string;
  sys_blood_pressure: string;
  oxygen_saturation: string;
  temperature: string;
  steps: string;
  distance: string;
  weight: string;
  email: string;
  z_accel: string;
  sex: string;
  address_street: string;
  address_number: string;
  address_city: string;
  address_postalcode: string;
}